Normalize accepted assets for places list
