+++ app/api/places/[id]/route.ts
@@ -3,15 +3,11 @@ import type { PoolClient } from "pg";
 
 import { getDbPool, hasDatabaseUrl } from "@/lib/db";
 import { places as fallbackPlaces } from "@/lib/data/places";
+import { normalizeAccepted, type PaymentAccept } from "@/lib/places/normalizeAccepted";
 
 const allowedVerificationLevels = ["owner", "community", "directory", "unverified"] as const;
 
 type Verification = (typeof allowedVerificationLevels)[number];
-
-type PaymentAccept = {
-  asset: string | null;
-  chain: string | null;
-};
 
 type Social = {
   platform: string;
@@ -94,45 +90,6 @@ const normalizeHours = (raw: string[] | string | null): string[] | null => {
     .filter(Boolean);
 
   return hours.length ? hours : null;
-};
-
-const normalizeAccepted = (payments: PaymentAccept[], fallback?: string[]): string[] => {
-  if (payments.length === 0) {
-    return fallback ?? [];
-  }
-
-  const normalized: string[] = [];
-  const seen = new Set<string>();
-
-  for (const payment of payments) {
-    const asset = payment.asset?.trim().toUpperCase() ?? null;
-    const chain = payment.chain?.trim().toUpperCase() ?? null;
-    let label: string | null = null;
-
-    if (
-      chain === "LIGHTNING" ||
-      chain === "LN" ||
-      asset === "LIGHTNING" ||
-      (asset === "BTC" && chain === "LIGHTNING")
-    ) {
-      label = "Lightning";
-    } else if (asset) {
-      label = asset;
-    } else if (chain) {
-      label = chain;
-    }
-
-    if (label && !seen.has(label)) {
-      seen.add(label);
-      normalized.push(label);
-    }
-  }
-
-  if (normalized.length === 0 && fallback?.length) {
-    return fallback;
-  }
-
-  return normalized;
 };
 
 const normalizeAbout = (value: string | null, verification: Verification) => {
